+++ tests/helpers.py
@@ -3,7 +3,20 @@ def raffler_params():
         "submission_url": "https://redd.it/4re9cx",
         "winner_count": 5,
         "min_account_age": 5,
+        "min_combined_karma": None,
         "min_comment_karma": 5,
         "min_link_karma": 5,
+        "ignored_users": ["TestUser"],
+    }
+
+
+def raffler_params_combined_karma():
+    return {
+        "submission_url": "https://redd.it/4re9cx",
+        "winner_count": 5,
+        "min_account_age": 5,
+        "min_combined_karma": 15,
+        "min_comment_karma": None,
+        "min_link_karma": None,
         "ignored_users": ["TestUser"],
     }